pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "switchboard"
version = "0.1.0"
description = "Policy-driven model selection service: clients → SwitchBoard → 9router → providers"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Velascat" },
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "respx>=0.21.0",
]

[project.scripts]
switchboard = "switchboard.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/switchboard"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["test"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
